requirement.txt
pyttsx3==2.90
gTTS==2.1.1
pyPDF2==1.26.0
pyPDF4==1.27.0
pdfrw==0.4
textract==1.6.3
pdfplumber
pdfminer.six==20200726
pytesseract==0.3.6
pdf2image==1.14.0
Pillow==7.2.0